Doing a slight visual update to the project. Moved the district finding algo to the welcome_controller.
I moved the CSS into a SCSS file so we can leverage SASS features.
Refactored application.html.erb into layouts that are rendered in.

Since the request is currently being handled in the welcome_controller,
I moved the district finding algo to the welcome_controller.

The district finding algo  doesn't seem to work, but I could be doing
something wrong. RGeo is parsing the geojson fine, but it's probably
related to how the latlngs are compared in pn_poly.